requirements.txt
asgiref==3.8.1
dj-database-url==3.0.0
Django==5.2.3
gunicorn==23.0.0
packaging==25.0
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.14.0
whitenoise==6.9.0
